+++ scripts/build.mjs
@@ -20,13 +20,13 @@ import postcssImport from 'postcss-import';
 
 import path from 'path';
 
-import po2js from './translations/po2js.mjs';
+import po2js from './translations/rollup-po2js/po2js.mjs';
 
 // Indicates whether to emit source maps
 const sourceMap = process.env.DEVELOPMENT || false;
 
 // Current working directory
-const cwd = path.resolve();
+const cwd = process.cwd();
 
 // Content of the `package.json`
 const pkg = JSON.parse( await readFile( path.join( cwd, 'package.json') ) );
@@ -95,8 +95,9 @@ export default [
 				sourceMap
 			} ),
 			po2js( {
-				sourceFolder: `${cwd}/lang/translations`,
-				destFolder: `${cwd}/dist/translations`,
+				type: 'single',
+				sourceFolder: path.join( cwd, 'lang', 'translations' ),
+				destFolder: path.join( cwd, 'dist', 'translations' ),
 				banner
 			} )
 		]
